Add `KPointSet` class to manage high-symmetry momentum points
This commit introduces the `KPointSet` class, which encapsulates named high-symmetry momentum points associated with a reciprocal lattice. The class provides methods for creating a point set from momentum values or fractional coordinates, rebasing points in a new reciprocal lattice, and resolving momentum by name. Additionally, the `KPointSet` is imported in the `geometries` module, while its previous references in the `symbolics` module have been removed to streamline the codebase.
